LayoutTests/platform/gtk/imported/w3c/web-platform-tests/mathml/relations/css-styling/overflow/computed-value-001-expected.txt


FAIL overflow on MathML elements assert_equals: overflow on mtr expected "scroll" but got "visible"
